Revert "unl0kr: store keyboard mode in a long instead of an int"
This reverts commit 4e13c312241420cbb3e5cc7d4f0dd3e5d17449be.

+++ unl0kr/terminal.c
@@ -24,7 +24,7 @@
 static int current_fd = -1;
 
 static int original_mode = -1;
-static long original_kb_mode = -1;
+static int original_kb_mode = -1;
 
 
 /**
